move GithubWatcher to Autoproj::Daemon
